pyproject.toml
[project]
name = "medical-diagnostic-agent"
version = "0.1.0"
description = "Medical diagnostic agent using Google ADK for symptom collection and assessment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.9.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
line_length = 88
profile = "black"

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]